Check for anaconda before instaling it

+++ installs/ubuntu_setup.sh
@@ -23,14 +23,18 @@ export RUNZSH=no
 sh -c "$(wget -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
 # Install anaconda
-ANACONDA_SH=Anaconda3-2020.07-Linux-x86_64.sh
-if [ ! -f ${ANACONDA_SH} ]; then
-	wget https://repo.anaconda.com/archive/${ANACONDA_SH}
-fi
-bash ./${ANACONDA_SH}
-rm ${ANACONDA_SH}
-if [ -d ~/anaconda3/ ]; then
-	~/anaconda3/bin/conda init zsh
+if [ ! -d ~/anaconda3/ ]; then
+    ANACONDA_SH=Anaconda3-2020.07-Linux-x86_64.sh
+    if [ ! -f ${ANACONDA_SH} ]; then
+        wget https://repo.anaconda.com/archive/${ANACONDA_SH}
+    fi
+    bash ./${ANACONDA_SH}
+    rm ${ANACONDA_SH}
+    if [ -d ~/anaconda3/ ]; then
+        ~/anaconda3/bin/conda init zsh
+    fi
+else
+    echo "Anaconda already installed; skipping installation..."
 fi
 
 # Copy configuration files to proper folders
@@ -51,3 +55,5 @@ git config --global core.editor "vim"
 git config --global diff.tool vimdiff
 git config --global merge.tool vimdiff
 
+# Done!
+echo "Done! Enjoy your new environment."
